% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORESTPLOT.PAC.R
\name{get_forestplot}
\alias{get_forestplot}
\title{基于 tidyverse 创建一个绘制森林图的函数，我真的服了现有的这些不符合我的需求的，零零散散只能差强人意的绘制森林图的R包了
这是在赶DDL的时候制作的R包，因此诸多部分在后续会被再次更新
由于语法简单，强烈建议在必要的时候，直接对本函数进行修改以及重编
复杂需求并不支持，复杂需求请自行使用tidyverse绘制
本函数的主要目的是，结合我们常用的，需要在PPT中完成后续的进一步调整的需求之上，快速稳定合理地输出森林图
并尽最大可能减少后续在PPT中的调整门槛和步骤}
\usage{
get_forestplot(
  df = NULL,
  left_side_data = NULL,
  right_side_data = NULL,
  estimate = NULL,
  ci_low = NULL,
  ci_high = NULL,
  xlimit = c(-1, 2),
  xstep = 0.5,
  ref_line_x = 0,
  ref_line_color = "black",
  ref_line_type = "dashed",
  right_mv = 0.75,
  first_col_mv = -0.75,
  last_col_mv = 0.2,
  chr_in_quote = " to ",
  box_shape = 15,
  box_size = 3,
  box_color = "black",
  headline_size = 0.5,
  headline_color = "black",
  x_line_size = 0.5,
  xline_color = "black",
  xline_text_size = 12,
  xline_text_color = "black",
  xline_text_family = "Arial",
  xline_text_bi = "plain",
  text_table_size = 4,
  text_table_color = "black",
  text_table_family = "Arial",
  text_table_bi = "plain",
  text_heading_size = 4,
  text_heading_family = "Arial",
  global_digits = 2
)
}
\arguments{
\item{df}{data_frame, 用于绘图的总数据集，必须提供，请保证与后续左右显示表格层数据集行数一致，本函数不提供数据集各式校验}

\item{left_side_data}{data_frame, 用于展示在左边的数据表格，必须提供，请保证与后续的置信区间行数一致，本函数不提供格式校验}

\item{right_side_data}{data_frame, 用于展示在右边的数据表格，可以缺省，本函数不提供格式校验}

\item{estimate}{vector, 效应值，请保证该效应在尺度或转换上与后续的置信区间一致，本函数不提供期望或自然对数转换}

\item{ci_low}{vector, 效应置信区间下限，请自行计算置信区间下限，本函数不提供se辅助计算}

\item{ci_high}{vector, 效应置信区间上限，请自行计算置信区间上限，本函数不提供se辅助计算}

\item{xlimit}{c(xmin, xmax), 规定x轴的最小值和最大值}

\item{xstep}{num, 规定x轴的刻度步进}

\item{ref_line_x}{num, 规定参考线的水平位置}

\item{ref_line_color}{string，规定参考线的颜色}

\item{ref_line_type}{string, 规定参考线的类型}

\item{right_mv}{num，指定一个表格层整体移动参数来调整表格和图形的相对位置}

\item{first_col_mv}{num, 指定一个表格层第一列移动参数来调整第一列和后续列的相对位置}

\item{last_col_mv}{num, 指定一个表格层最后一列移动参数来调整最后一列列和前列的相对位置}

\item{chr_in_quote}{string，指定一个自动置信区间中的分割符号，默认为' to '}

\item{box_shape}{num，指定图形中间点的形状}

\item{box_size}{num，指定图形中间点的大小}

\item{box_color}{string, 指定图形中间点的颜色，本函数不提供fill参数}

\item{headline_size}{num, 指定纵标目下边线的粗细}

\item{headline_color}{string，指定纵标目下边线的颜色}

\item{x_line_size}{num，指定x轴线的粗细}

\item{xline_color}{string，指定x轴线的颜色}

\item{xline_text_size}{string，指定x轴字体}

\item{xline_text_color}{string，指定x轴字体}

\item{xline_text_family}{string，指定x轴字体}

\item{xline_text_bi}{string，指定x轴是否需要加粗或斜体}

\item{text_table_size}{num, 指定图形表格层的文字大小}

\item{text_table_color}{string，指定图形表格层的文字颜色}

\item{text_table_family}{string, 指定图形表格层的文字字体，请确保字体可用，本函数不做字体可用性以及UTF-8校验}

\item{text_table_bi}{string，指定图形表格层文字是否需要加粗或者斜体}

\item{text_heading_size}{num，指定图形纵标目字体大小}

\item{text_heading_family}{string, 指定图形纵标目的文字字体，请确保字体可用，本函数不做字体可用性以及UTF-8校验}

\item{global_digits}{int, 指定图形层全局的数字小数位数，不足则强制补0}
}
\value{
A ggplot2 object
}
\description{
基于 tidyverse 创建一个绘制森林图的函数，我真的服了现有的这些不符合我的需求的，零零散散只能差强人意的绘制森林图的R包了
这是在赶DDL的时候制作的R包，因此诸多部分在后续会被再次更新
由于语法简单，强烈建议在必要的时候，直接对本函数进行修改以及重编
复杂需求并不支持，复杂需求请自行使用tidyverse绘制
本函数的主要目的是，结合我们常用的，需要在PPT中完成后续的进一步调整的需求之上，快速稳定合理地输出森林图
并尽最大可能减少后续在PPT中的调整门槛和步骤
}
